Creating a file for OptionManagerMixin and helper functions
